requirements.txt
matplotlib==3.7.4
numpy==1.24.4
onnxruntime==1.20.1
opencv_python==4.11.0.86
opencv_python_headless==4.5.5.64
Pillow==10.1.0
Pillow==11.1.0
pycocotools==2.0.8
PyYAML==6.0.2
PyYAML==6.0.2
safetensors==0.5.3
scipy==1.15.2
setuptools==72.1.0
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
timm==0.9.10
torch==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.66.6
